Migrate assertions to Kotest
